+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## v14.0.0 - 2018-11-13
+
+Complete re-work of network module to provide the ability to vary network infra.
+
+## v13.1.0 - 2018-11-12
+
+Allow custom domains to be created seperately from stages.
+
 ## v13.0.0 - 2018-11-10
 
 Refactoring API GW module for further flexibility